Доработаны функции
