+++ js/main.js
@@ -1,10 +1,11 @@
-import {generateImagesFromData} from './image-generator.js';
-import {getData} from './network.js';
+import {renderImagesFromData} from './image-render.js';
+import {getData} from './data.js';
 import {showAlert} from './util.js';
-import {addFilterListeners} from './image-generator.js';
+import {addFilterListeners, applyDefaultFilter, showFilterMenu} from './filter.js';
 import './image-upload.js';
 
 getData((imageDescriptions) => {
-  generateImagesFromData(imageDescriptions);
-  addFilterListeners(imageDescriptions);
+  showFilterMenu();
+  applyDefaultFilter(imageDescriptions, () => renderImagesFromData(imageDescriptions));
+  addFilterListeners(imageDescriptions, renderImagesFromData);
 }, showAlert);
